# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Education IS Swagger OpenAPI documentation",
        "description": "OpenAPI documentation for backend methods of Education IS.<br>Notice: every data figuring in response is placing in <i>Result</i> Schema (see in description below).",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8002",
            "description": "DEV"
        },
        {
            "url": "http://localhost:8002",
            "description": "PROD"
        }
    ],
    "paths": {
...
        "/api/v1/email/verify/{id}/{hash}/": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Route to verify user's email.",
                "operationId": "e4f75b2b6caa022a261dde4883a9ea60",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    },
                    {
                        "name": "hash",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Result"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "There is errors in request.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Result"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Result": {
                "description": "Server answer",
                "properties": {
                    "Success": {
                        "type": "boolean",
                        "default": "false"
                    },
                    "Content": {
                        "description": "Response content in case of successful request. Can be an object or an array if multiple object returns (list of smth).",
                        "type": "object",
                        "items": []
                    },
                    "Error": {
                        "description": "Error message if Success flag is false.",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}